+++ gems/aws-sdk-lookoutmetrics/CHANGELOG.md
@@ -1,6 +1,11 @@
 Unreleased Changes
 ------------------
 
+1.4.0 (2021-06-14)
+------------------
+
+* Feature - Added "LEARNING" status for anomaly detector and updated description for "Offset" parameter in MetricSet APIs.
+
 1.3.0 (2021-05-28)
 ------------------